pyproject.toml
[tool.poetry]
name = "document-scanner"
version = "0.1.0"
description = "OCR application for scanning identity documents"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "4.6.0.66"
pytesseract = "^0.3.10"
Pillow = "^10.2.0"
numpy = "<2.0.0"
paddlepaddle = "2.5.1"
albumentations = "1.3.1"
paddleocr = "2.7.0"
transformers = "^4.46.2"
torch = "^2.5.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 